// src/main/webapp/app/impact-evaluation/model/enum/attack-cost-param-type.enum.ts
export enum AttackCostParamType {
    NUMBER_OF_CUSTOMERS = 0,
    PROTECTION_COST_PER_CUSTOMER = 1,
    NOTIFICATION_COST_PER_CUSTOMER = 2,
    EMPLOYEE_COST_PER_HOUR = 3,
    FRACTION_OF_EMPLOYEES_AFFECTED_BY_OUTAGE = 4,
    AVERAGE_REVENUE_PER_HOUR = 5,
    FRACTION_OF_REVENUE_AFFECTED_BY_OUTAGE = 6,
    RECOVERY_COST = 7,
    /*---*/REPAIR_SERVICES = 8,
    /*---*/REPLACEMENT_PARTS = 9,
    /*---*/LOST_DATA_RECOVERY = 10,
    /*---*/OTHER_COSTS = 11
}